// apps/Conduit-UI/src/models/notifications/NotificationModels.ts
export interface NotificationData {
  userIds: string[];
  title: string;
  body: string;
}

export interface INotificationConfig {
  active: boolean;
  providerName: string;
  firebase?: FirebaseSettings;
  onesignal?: OneSignalSettings;
  message?: string;
}

export enum INotificationProviders {
  firebase = 'firebase',
  onesignal = 'onesignal',
}

export interface INotificationProviderSettings {
  firebase: FirebaseSettings;
  onesignal: OneSignalSettings;
}

export interface FirebaseSettings {
  projectId: string;
  privateKey: string;
  clientEmail: string;
}

export interface OneSignalSettings {
  appId: string;
  apiKey: string;
}